feat: categorize StoryScenes
