+++ mc1-aidle.swiftpm/Model/StoryScene/EndingStoryScene.swift
@@ -1,0 +1,11 @@
+//
+//  EndingStoryScene.swift
+//  mc1-aidle
+//
+//  Created by byo on 2023/03/28.
+//
+
+import Foundation
+
+final class EndingStoryScene: StoryScene {
+}
